+++ Core/include/Acts/TrackFitting/GainMatrixUpdater.hpp
@@ -12,8 +12,8 @@
 #include "Acts/EventData/MeasurementHelpers.hpp"
 #include "Acts/EventData/MultiTrajectory.hpp"
 #include "Acts/EventData/TrackParameters.hpp"
-#include "Acts/Fitter/KalmanFitterError.hpp"
-#include "Acts/Fitter/detail/VoidKalmanComponents.hpp"
+#include "Acts/TrackFitting/KalmanFitterError.hpp"
+#include "Acts/TrackFitting/detail/VoidKalmanComponents.hpp"
 #include "Acts/Utilities/Definitions.hpp"
 #include "Acts/Utilities/Logger.hpp"
 #include "Acts/Utilities/Result.hpp"
