+++ Tests/UnitTests/Core/CMakeLists.txt
@@ -1,14 +1,14 @@
 add_unittest(Version VersionTests.cpp)
 
 add_subdirectory(EventData)
-add_subdirectory(Fitter)
 add_subdirectory(Geometry)
 add_subdirectory(MagneticField)
 add_subdirectory(Material)
 add_subdirectory(Propagator)
 add_subdirectory(Seeding)
 add_subdirectory(Surfaces)
+add_subdirectory(TrackFinding)
+add_subdirectory(TrackFitting)
 add_subdirectory(Utilities)
 add_subdirectory(Vertexing)
 add_subdirectory(Visualization)
-add_subdirectory(TrackFinder)
