+++ Tests/UnitTests/Core/TrackFitting/GainMatrixSmootherTests.cpp
@@ -10,9 +10,9 @@
 
 #include "Acts/EventData/Measurement.hpp"
 #include "Acts/EventData/MeasurementHelpers.hpp"
-#include "Acts/Fitter/GainMatrixSmoother.hpp"
 #include "Acts/Surfaces/CylinderSurface.hpp"
 #include "Acts/Tests/CommonHelpers/FloatComparisons.hpp"
+#include "Acts/TrackFitting/GainMatrixSmoother.hpp"
 #include "Acts/Utilities/ParameterDefinitions.hpp"
 
 #include <memory>
